feat: AttractionRepository 클래스 PK 재설정

+++ src/main/java/com/ssafy/tourmates/admin/attraction/repository/AttractionRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.tourmates.admin.attraction.repository;
 import com.ssafy.tourmates.admin.attraction.AttractionInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttractionRepository extends JpaRepository<AttractionInfo, Long>, AttractionRepositoryCustom {
+public interface AttractionRepository extends JpaRepository<AttractionInfo, Integer>, AttractionRepositoryCustom {
 }
